Prevent IDOR vulnerability in PUT, PATCH and DELETE endpoints
Before this commit, one can simply send a request to the endpoint with a
modified to-do ID, and the server will go ahead and perform the
requested action without any authorization checks (i.e., attackers can
modify the to-dos of other users, a.k.a. Insecure direct object
references). This commit fixes that by explicitly checking whether the
user ID associated with the to-do matches the user ID of the
authenticated user first before taking any action in the PUT, PATCH and
DELETE endpoints.