Change database to PostgreSQL
The SQLite backend does not support RETURNING clauses, so calling
`get_result` on `INSERT` / `UPDATE` / `DELETE` statements are not
possible (which I need to prevent a second call to the `GET /api/todos`
endpoint, which is wasteful). So, I changed the database to PostgreSQL.

Some modifications were also made to the database schema: the ID type
has changed from SERIAL to UUID, and some fields were marked NOT NULL so
unnecessary Nullable types won't appear in schemas.rs.

Before this commit, every time a to-do was created, edited, deleted, or
has its status updated, a second call to the `GET /api/todos` endpoint
was issued to retrieve the updated list of to-dos for presentation by
Svelte. However, this can be quite expensive if the number of to-dos is
large, so by opting to return only the relevant information in the
response of the POST / PUT / PATCH / DELETE endpoints and with the help
of PostgreSQL (since it supports RETURNING clauses, unlike SQLite), we
can avoid the second call to the `GET /api/todos` endpoint.

Now:

- The POST endpoint (create to-do) will return the newly-created to-do
    in the response body.
- The PUT endpoint (update to-do) will return the updated to-do in the
    response body.
- The PATCH endpoint (update to-do status) will return the updated to-do
    status in the response body.
- The DELETE endpoint (delete to-do) will return the deleted to-do's ID
    in the response body.

To facilitate the change to PostgreSQL, the character length limits for
the to-do title and contents are now strictly enforced (since this limit
is not strictly enforced by SQLite). I added more validation checks when
destructuring into the endpoint data structs.