Change the API for creating messages
Previously, the `create_message` function was used to create messages.
Now, it is done by calling the `create_message` trait method on the
message cookie, which looks cleaner. Also, the types of the message
category, title, and content were changed from `String` to `&str`, which
means there is no need to include redundant `.to_string()`s when
creating messages anymore.